Moved functionality out of main file and into seperate files for ease of editting and view.
