+++ lib/family.dart
@@ -1,0 +1,9 @@
+class Family {
+  String name;
+  String invited;
+
+  Family(String n, String i) {
+    name = n;
+    invited = i;
+  }
+}
