+++ lib/to-do
@@ -8,4 +8,10 @@ see about moving functions out to seperate class?
 
 clear all vars on sign out
 
+Have loading bounce back eventually
+
+In Enteremail maybe set firstRun later? After signin works.
+
 LOGIN VIEW
+
+fix login
